+++ deirokay/airflow/deirokay_operator.py
@@ -1,8 +1,9 @@
 import logging
+import warnings
 from datetime import datetime
 from typing import Optional, Union
 
-from airflow.exceptions import AirflowFailException, AirflowSkipException
+from airflow.exceptions import AirflowSkipException
 from airflow.models.baseoperator import BaseOperator
 
 import deirokay
@@ -16,13 +17,13 @@ logger = logging.getLogger(__name__)
 class DeirokayOperator(BaseOperator):
     """Parse a DataFrame from a file using Deirokay options and
     validate against a Deirokay Validation Document.
-    You may choose different severity levels to trigger a task
+    You may choose different severity levels to mark a task as
     "soft failure" (`skipped` state) or "normal failure" (`failed`
     state).
 
     Parameters
     ----------
-    path_to_file : str
+    data / path_to_file (deprecated) : str
         File to be parsed into Deirokay.
     options : Union[dict, str]
         A dict or a local/S3 path to a YAML/JSON options file.
@@ -38,48 +39,75 @@ class DeirokayOperator(BaseOperator):
         Minimum Deirokay severity level to trigger a
         "soft failure".
         Any statement with lower severity level will only raise a
-        warning.
+        warning. Set to `None` to never trigger.
         By default SeverityLevel.MINIMAL (1).
     hard_fail_level : int, optional
         Minimum Deirokay severity level to trigger a task failure.
+        Set to `None` to never trigger.
         By default SeverityLevel.CRITICAL (5).
+    reader_kwargs : dict, optional
+        Additional keyword arguments for `Deirokay.data_reader` method.
+    validator_kwargs : dict, optional
+        Additional keyword arguments for `Deirokay.validate` method.
     """
 
     template_fields = [
-        'path_to_file',
+        'data',
         'options',
         'against',
         'template',
-        'save_to'
+        'save_to',
+        'reader_kwargs',
+        'validator_kwargs'
     ]
     template_fields_renderers = {'options': 'json', 'against': 'json'}
     ui_color = '#59f75e'
 
     def __init__(
         self,
-        path_to_file: str,
-        options: Union[dict, str],
-        against: Union[dict, str],
+        data: str = None,
+        path_to_file: str = None,
+        options: Union[dict, str] = None,
+        against: Union[dict, str] = None,
         template: Optional[dict] = None,
         save_to: Optional[str] = None,
         soft_fail_level: int = SeverityLevel.MINIMAL,
         hard_fail_level: int = SeverityLevel.CRITICAL,
+        reader_kwargs: Optional[dict] = None,
+        validator_kwargs: Optional[dict] = None,
         **kwargs
     ):
         super().__init__(**kwargs)
 
-        self.path_to_file = path_to_file
+        assert bool(data) is not bool(path_to_file), (
+            'Declare either `data` or `path_to_file`, but not both.'
+        )
+        if path_to_file:
+            warnings.warn(
+                'The argument `path_to_file` is deprecated and will be'
+                ' removed in next major release. Use `data` instead.',
+                DeprecationWarning
+            )
+        assert options
+        assert against
+        self.data = data or path_to_file
         self.options = options
         self.against = against
         self.template = template
         self.save_to = save_to
         self.soft_fail_level = soft_fail_level
         self.hard_fail_level = hard_fail_level
+        self.reader_kwargs = reader_kwargs or {}
+        self.validator_kwargs = validator_kwargs or {}
 
     # docstr-coverage:inherited
     def execute(self, context: dict):
         current_date = datetime.strptime(context['ts_nodash'], '%Y%m%dT%H%M%S')
-        df = deirokay.data_reader(self.path_to_file, options=self.options)
+        df = deirokay.data_reader(
+            self.data,
+            options=self.options,
+            **self.reader_kwargs
+        )
 
         validation_document = deirokay.validate(
             df,
@@ -87,12 +115,19 @@ class DeirokayOperator(BaseOperator):
             template=self.template,
             save_to=self.save_to,
             current_date=current_date,
-            raise_exception=False
+            raise_exception=False,
+            **self.validator_kwargs
         )
         try:
             raise_validation(validation_document, SeverityLevel.MINIMAL)
         except ValidationError as e:
-            if e.level >= self.hard_fail_level:
-                raise AirflowFailException from e
-            if e.level >= self.soft_fail_level:
+            if (
+                self.hard_fail_level is not None and
+                e.level >= self.hard_fail_level
+            ):
+                raise e
+            if (
+                self.soft_fail_level is not None and
+                e.level >= self.soft_fail_level
+            ):
                 raise AirflowSkipException from e