+++ tests/statements/test_not_null.py
@@ -1,4 +1,6 @@
-from deirokay import data_reader, validate
+from pprint import pprint
+
+from deirokay import data_reader, profile, validate
 
 
 def test_not_null_statement():
@@ -32,3 +34,23 @@ def test_not_null_statement():
     }
 
     validate(df, against=assertions)
+
+
+def test_profile_wont_generate_useless_statement():
+    """Prevent statement generation when `at_least_%` is 0."""
+    df = data_reader(
+        'tests/transactions_sample.csv',
+        options='tests/options.json'
+    )
+
+    validation_doc = profile(df, 'sample')
+    pprint(validation_doc)
+
+    # Check that the statement was not generated for "EMPTY" column
+    scope_empty = next(filter(
+        lambda x: x['scope'] == 'EMPTY', validation_doc['items']
+    ))
+    assert len([
+        stmt for stmt in scope_empty['statements']
+        if stmt['type'] == 'not_null'
+    ]) == 0